+++ src/plugins/replication.ts
@@ -14,7 +14,9 @@ import {
 } from 'rxjs';
 import {
     skipUntil,
-    filter
+    filter,
+    first,
+    mergeMap
 } from 'rxjs/operators';
 
 import {
@@ -36,11 +38,11 @@ import {
     isInstanceOf as isRxCollection
 } from '../rx-collection';
 import type {
-    RxQuery,
     RxCollection,
     PouchSyncHandler,
     PouchReplicationOptions,
-    RxPlugin
+    RxPlugin,
+    SyncOptions
 } from '../types';
 import { RxDBWatchForChangesPlugin } from './watch-for-changes';
 
@@ -50,20 +52,13 @@ addRxPlugin(PouchReplicationPlugin);
 // add the watch-for-changes-plugin
 addRxPlugin(RxDBWatchForChangesPlugin);
 
+/**
+ * Contains all pouchdb instances that
+ * are used inside of RxDB by collections or databases.
+ * Used to ensure the remote of a replication cannot be an internal pouchdb.
+ */
 const INTERNAL_POUCHDBS = new WeakSet();
 
-
-export interface SyncOptions {
-    remote: string | any;
-    waitForLeadership?: boolean;
-    direction?: {
-        push?: boolean,
-        pull?: boolean
-    };
-    // for options see https://pouchdb.com/api.html#replication
-    options?: PouchReplicationOptions;
-    query?: RxQuery;
-}
 export class RxReplicationStateBase {
     public _subs: Subscription[] = [];
 
@@ -83,8 +78,8 @@ export class RxReplicationStateBase {
     public canceled: boolean = false;
 
     constructor(
-        public collection: RxCollection,
-        private syncOptions: SyncOptions
+        public readonly collection: RxCollection,
+        public readonly syncOptions: SyncOptions
     ) {
         // create getters
         Object.keys(this._subjects).forEach(key => {
@@ -118,15 +113,20 @@ export class RxReplicationStateBase {
         );
     }
 
-    cancel() {
+    /**
+     * Returns false when the replication has already been canceled
+     */
+    cancel(): Promise<boolean> {
         if (this.canceled) {
-            return;
+            return Promise.resolve(false);
         }
         this.canceled = true;
         if (this._pouchEventEmitterObject) {
             this._pouchEventEmitterObject.cancel();
         }
         this._subs.forEach(sub => sub.unsubscribe());
+
+        return Promise.resolve(true);
     }
 }
 
@@ -207,7 +207,23 @@ export function setPouchEventEmitter(
                 Promise.all(unhandledEvents).then(() => rxRepState._subjects.complete.next(info));
             })
     );
-
+    // auto-cancel one-time replications on complelete to not cause memory leak
+    if (
+        !rxRepState.syncOptions.options ||
+        !rxRepState.syncOptions.options.live
+    ) {
+        rxRepState._subs.push(
+            rxRepState.complete$.pipe(
+                filter(x => !!x),
+                first(),
+                mergeMap(() => {
+                    return rxRepState.collection.database
+                        .requestIdlePromise()
+                        .then(() => rxRepState.cancel());
+                })
+            ).subscribe()
+        );
+    }
 
     function getIsAlive(emitter: any): Promise<boolean> {
         // "state" will live in emitter.state if single direction replication